% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate an HTML Report}
\usage{
generateReport(
  folder = NULL,
  fileType = NULL,
  assay = NULL,
  plateOrder = NULL,
  title = "QC Report",
  output_file = "report.html",
  scale = TRUE
)
}
\arguments{
\item{folder}{Either a character for the file path or a dataElement to be included in the report.}

\item{fileType}{Character "xml", "tsv" or "txt" if supplying file path to the folder argument.}

\item{assay}{Character of "AA", "TRY", "BA" or "SCFA".}

\item{plateOrder}{Character vector of order plates should appear in. For
example c("COVp005", "COVp002", "COVp001"). The default is NULL resulting in
an order from Acquisition date (if present for all plates), or name order.}

\item{title}{The title of the report.}

\item{output_file}{The name of the output HTML file.}

\item{scale}{\itemize{
\item Boolean. TRUE if the data should be log10(value + 1)
}}
}
\value{
The path to the generated HTML report.
}
\description{
This function generates an HTML report using an R Markdown template.
}
