% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAAdae.R
\name{makeAAdae}
\alias{makeAAdae}
\title{makeAAdae}
\usage{
makeAAdae(result = NULL, file = NULL)
}
\arguments{
\item{result}{results from SQLite database}

\item{file}{Path name to folder containing files or path name to file.}
}
\value{
data element
}
\description{
makeAAdae
}
