% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{plotQC}
\usage{
plotQC(dae = NULL, scale = TRUE, plateOrder = NULL, optns = list())
}
\arguments{
\item{dae}{\itemize{
\item dae of targeted Mass Spectrometry.
}}

\item{scale}{\itemize{
\item Boolean. TRUE if the data should be log10(value + 1)
}}

\item{plateOrder}{NULL or character vector of order of plates desired.
Default is NULL. Supply as c("COVp002", "COVp001", "COVp020") etc.}

\item{optns}{list of options}
}
\value{
list of plots with a slot for each analyte.
}
\description{
Plots sample types across plates, excluding blanks. If Acquisition Date is
available and there are no NAs, plates appear in date order, otherwise in name
order or order provided (see plateOrder argument).
}
