% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryData.R
\name{summaryData}
\alias{summaryData}
\title{summaryData}
\usage{
summaryData(dae = NULL)
}
\arguments{
\item{dae}{DataElement}

\item{file}{path to a folder of multiple files or path to a single file.}
}
\value{
list of data per plate
}
\description{
Calculates the data for summaryTables and PassFail used to propagate colouring
in plotQC.
}
