% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTables.R
\name{summaryTables}
\alias{summaryTables}
\title{summaryTables}
\usage{
summaryTables(dae = NULL, plateOrder = NULL)
}
\arguments{
\item{dae}{\itemize{
\item dae of targeted Mass Spectrometry.
}}

\item{plateOrder}{Character vector of order plates should appear in. For
example c("COVp005", "COVp002", "COVp001"). The default is NULL resulting in
an order from Acquisition date (if present for all plates), or name order.}
}
\value{
list of tables and matching data with a slot for each plate. Criteria
for PASS, must meet all 3 criteria of R2 is ≥ 0.99, LTR RSD is ≤ 20 or "N/A"
(uses ltr sampleType Quantity), totalQCInjectionPass fraction of qc samples
with 80=< Accuracy/Recovery\verb{[\%]} =< 120 totalQCPass(\%), percent of qc injections
that have 80=< Accuracy/Recovery\verb{[\%]} =< 120. Pass needs to be ≥ 66\% (uses qc sampleType)
}
\description{
Displays status (pass or caution) for each analyte in each plate in table format.
}
