% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate an HTML Report}
\usage{
generateReport(
  data = NULL,
  file = NULL,
  title = "QC Report",
  output_file = "report.html",
  scale = TRUE
)
}
\arguments{
\item{data}{The dataset (dataElement) to be included in the report.}

\item{title}{The title of the report.}

\item{output_file}{The name of the output HTML file.}
}
\value{
The path to the generated HTML report.
}
\description{
This function generates an HTML report using an R Markdown template.
}
\examples{
#generateReport(data = bio, output_file = "~/Desktop/report.html")
}
