% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{plotQC}
\usage{
plotQC(dae, scale = TRUE, optns = list())
}
\arguments{
\item{dae}{\itemize{
\item dae of targeted Mass Spectrometry.
}}

\item{scale}{\itemize{
\item Boolean. TRUE if the data should be log10(value + 1)
}}

\item{optns}{list of options}
}
\value{
list of plots with a slot for each analyte.
}
\description{
Plots sample types across plates, excluding blanks.
}
