% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{plotQC}
\usage{
plotQC(dae, optn = list())
}
\arguments{
\item{dae}{\itemize{
\item dae of targeted Mass Spectrometry.
}}

\item{optns}{list of options
\itemize{
\item scale - Boolean. TRUE if the data should be log10(value + 1)
}}
}
\value{
list of plots with a slot for each analyte.
}
\description{
Plots sample types across plates, excluding blanks.
}
