% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryData.R
\name{summaryData}
\alias{summaryData}
\title{summarData}
\usage{
summaryData(dae = "", file = "")
}
\arguments{
\item{dae}{DataElement}
}
\value{
list of data per plate
}
\description{
Calculates the data for summaryTables and PassFail used to propagate colouring
in plotQC.
}
