% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTables.R
\name{summaryTables}
\alias{summaryTables}
\title{summaryTables}
\usage{
summaryTables(dae)
}
\arguments{
\item{dae}{\itemize{
\item dae of targeted Mass Spectrometry.
}}
}
\value{
list of tables and matching data with a slot for each plate. Criteria
for PASS, must meet all 3 criteria of R2 is ≥ 0.99, LTR RSD is ≤ 20 or "N/A"
(uses ltr sampleType Quantity), totalQCInjectionPass fraction of qc samples
with 80=< Accuracy/Recovery\verb{[\%]} =< 120 totalQCPass(\%), percent of qc injections
that have 80=< Accuracy/Recovery\verb{[\%]} =< 120. Pass needs to be ≥ 66\% (uses qc sampleType)
}
\description{
Displays status (pass or caution) for each analyte in each plate in table format.
}
